backend/requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
google-generativeai>=0.3.0,<1.0.0
numpy>=1.24.0,<2.0.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
openpyxl>=3.1.0,<4.0.0
pandas>=2.0.0,<3.0.0
pdfplumber>=0.10.0,<1.0.0
sse-starlette>=1.6.0,<3.0.0
slowapi>=0.1.9,<1.0.0
defusedxml>=0.7.0,<1.0.0
